Refactored parsing into the UIHelpers class
