+++ Widgets/WidgetBase/VarmintWidgetSerialization.cs
@@ -92,65 +92,6 @@ namespace MonoVarmint.Widgets
             }
         }
 
-        //--------------------------------------------------------------------------------------
-        /// <summary>
-        /// ParseVector
-        /// </summary>
-        //--------------------------------------------------------------------------------------
-        public static Vector2 ParseVector(string text)
-        {
-            var parts = text.Split(',');
-            if (parts.Length != 2)
-            {
-                throw new ApplicationException("Bad Vector Specification");
-            }
-            return new Vector2(float.Parse(parts[0].Trim()), float.Parse(parts[1].Trim()));
-        }
-
-        //--------------------------------------------------------------------------------------
-        /// <summary>
-        /// Point
-        /// </summary>
-        //--------------------------------------------------------------------------------------
-        public static Point ParsePoint(string text)
-        {
-            var parts = text.Split(',');
-            if (parts.Length != 2)
-            {
-                throw new ApplicationException("Bad Point Specification");
-            }
-            return new Point(int.Parse(parts[0].Trim()), int.Parse(parts[1].Trim()));
-        }
-
-        //--------------------------------------------------------------------------------------
-        /// <summary>
-        /// ParseColor
-        /// </summary>
-        //--------------------------------------------------------------------------------------
-        public static Color ParseColor(string text)
-        {
-            // #ARGB - formatted
-            if (text.StartsWith("#"))
-            {
-                var rawColor = Convert.ToUInt32(text.Substring(1), 16);
-                var blue = (int)(rawColor & 0xff);
-                rawColor >>= 8;
-                var green = (int)(rawColor & 0xff);
-                rawColor >>= 8;
-                var red = (int)(rawColor & 0xff);
-                rawColor >>= 8;
-                var alpha = (int)(rawColor & 0xff);
-                if (text.Length <= 7) alpha = 255;
-                return Color.FromNonPremultiplied(red, green, blue, alpha);
-            }
-
-            // Color name formatted
-            var colorType = typeof(Color);
-            var colorProperty = colorType.GetProperty(text);
-            if (colorProperty == null) throw new ApplicationException("Unknown color: " + text);
-            return (Color)colorProperty.GetValue(null, null);
-        }
-
         Dictionary<string, string> _bindingTemplates = new Dictionary<string, string>();
         BindingFlags _publicInstance = BindingFlags.Public | BindingFlags.Instance;
         //--------------------------------------------------------------------------------------
@@ -212,8 +153,8 @@ namespace MonoVarmint.Widgets
                     var valueParts = keyValuePair.Split(new char[] { '=' }, 2);
                     if (valueParts.Length != 2) throw new ApplicationException("Expected format of Name=Value");
                     var types = propertyInfo.PropertyType.GenericTypeArguments;
-                    var key = GetValueFromText(types[0], valueParts[0]);
-                    var value = GetValueFromText(types[1], valueParts[1]);
+                    var key = UIHelpers.GetValueFromText(types[0], valueParts[0]);
+                    var value = UIHelpers.GetValueFromText(types[1], valueParts[1]);
 
                     // Create dictionary object if not there
                     var dictionary = propertyInfo.GetValue(this);
@@ -230,40 +171,11 @@ namespace MonoVarmint.Widgets
             }
             else
             {
-                propertyInfo.SetValue(this, GetValueFromText(propertyInfo.PropertyType, valueText));
+                propertyInfo.SetValue(this, UIHelpers.GetValueFromText(propertyInfo.PropertyType, valueText));
             }
         }
 
-        //--------------------------------------------------------------------------------------
-        /// <summary>
-        /// GetValueFromText - Convert text from a vwml file into an value
-        /// </summary>
-        //--------------------------------------------------------------------------------------
-        private object GetValueFromText(Type type, string valueText)
-        {
-            switch (type.Name)
-            {
-                case "String": return valueText;
-                case "Vector2": return ParseVector(valueText);
-                case "Point": return ParsePoint(valueText);
-                case "Single":
-                case "float": return float.Parse(valueText);
-                case "double": return double.Parse(valueText);
-                case "Int32":
-                case "int": return int.Parse(valueText);
-                case "Int64":
-                case "long": return long.Parse(valueText);
-                case "Boolean":
-                case "bool": return Boolean.Parse(valueText);
-                case "Color": return ParseColor(valueText);
-                default:
-                    if (type.IsEnum) return Enum.Parse(type, valueText);
-                    else if (type.Name == "Object") return valueText;
-                    else if (type.IsClass) return Activator.CreateInstance(type, valueText);
-                    else throw new ApplicationException("Don't know create a " + type.Name);
-            }
-        }
-
+        
         //--------------------------------------------------------------------------------------
         /// <summary>
         /// IsBinding - true if the text is a binding specifier
